批次建機器：階段 5 的最後一塊
補連線的提案會說「目標的服務 redis 在 dev 一台都還沒建，要先建機器」。
沒有這個功能，那句話就是死路——工具叫你去做一件它不讓你做的事。

同樣是 L001，subject 是什麼決定了要怎麼修：
  Relationship → 補一條連線
  Container    → 批次建機器
  外部系統     → 還沒做
這個分辨是規則，所以在 fix_for 做完，前端只是照著開對應的表單。

batch.rs 的引擎早就寫好了，這次補上：
- batch::plan 檢查撞名與接點歸屬，並且把 id 發好（apply 因此是決定性的）
- 撞名一律擋下來不是警告：同一個環境兩台 redis-01 會讓萬用字元數到 2，
  而使用者以為那是兩台不同的機器——正好是這工具要防的誤會
- 接點不屬於那個服務也擋：放過去的話 lint 會變成一個很難懂的 L003
- Edit::AddInstances，可以掛在站點底下，可以整批復原

BatchSpec 的數字從 usize 改成 u32（要建幾台不是記憶體索引，
而且 usize 匯不出 TypeScript）。

畫面每改一個字就重算預覽——一次建六台是會後悔的操作，
按下去之前就要看到那六行。送出用的是預覽算好的那批節點，不重算。

建完機器再補連線，整個環境就乾淨了：
tests/add_instances.rs::建完機器再補連線_整個環境就乾淨了 走完整圈。

Rust 236 測試、前端 85 測試。

+++ crates/loom-core/tests/batch_into_lint.rs
@@ -15,7 +15,7 @@ use loom_core::logical::Protocol;
 
 /// 生出 `count` 台 Redis 的規格。名稱與位址的起始值刻意錯開，
 /// 對應現實中 `redis-01`～`redis-12` 配 `10.0.1.11`～`10.0.1.22` 的情形。
-fn redis_batch(count: usize) -> BatchSpec {
+fn redis_batch(count: u32) -> BatchSpec {
     BatchSpec {
         count,
         name_template: "redis-{n}".into(),
@@ -35,7 +35,7 @@ fn redis_batch(count: usize) -> BatchSpec {
 }
 
 /// 用批次建立換掉 prod 原本手寫的 Redis，並把連線的 expect 設成 `expect`。
-fn prod_with_batch(count: usize, expect: u32) -> loom_core::Project {
+fn prod_with_batch(count: u32, expect: u32) -> loom_core::Project {
     let mut project = healthy_project();
     let prod = &mut project.environments[0];
